Add Secret validation

+++ src/main/java/com/sap/subsystem/vcs_repository/facade/VcsRepositoryFacade.java
@@ -1,6 +1,9 @@
 package com.sap.subsystem.vcs_repository.facade;
 
 import com.sap.subsystem.common.error.exception.DuplicateEntityException;
+import com.sap.subsystem.common.error.exception.GithubException;
+import com.sap.subsystem.common.error.exception.InvalidSecretException;
+import com.sap.subsystem.github_api.domain.dto.GithubRepositoryApiResponseDto;
 import com.sap.subsystem.github_api.service.GithubRepositoryApiService;
 import com.sap.subsystem.vcs_repository.domain.dto.EditVcsRepositoryDto;
 import com.sap.subsystem.vcs_repository.domain.dto.VcsRepositoryDto;
@@ -9,6 +12,8 @@ import com.sap.subsystem.vcs_repository.domain.mapping.VcsRepositoryMapper;
 import com.sap.subsystem.vcs_repository.domain.model.VcsRepository;
 import com.sap.subsystem.vcs_repository.service.VcsRepositoryService;
 import org.apache.commons.lang3.ObjectUtils;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,12 +23,13 @@ import java.util.UUID;
 @Component
 @Transactional(readOnly = true)
 public class VcsRepositoryFacade {
+
     private final VcsRepositoryService vcsRepositoryService;
     private final VcsRepositoryMapper vcsRepositoryMapper;
     private final GithubRepositoryApiService githubRepositoryApiService;
 
-    public VcsRepositoryFacade(VcsRepositoryService vcsRepositoryService, VcsRepositoryMapper vcsRepositoryMapper,
-                               GithubRepositoryApiService githubRepositoryApiService) {
+    public VcsRepositoryFacade(final VcsRepositoryService vcsRepositoryService, final VcsRepositoryMapper vcsRepositoryMapper,
+                               final GithubRepositoryApiService githubRepositoryApiService) {
         this.vcsRepositoryService = vcsRepositoryService;
         this.vcsRepositoryMapper = vcsRepositoryMapper;
         this.githubRepositoryApiService = githubRepositoryApiService;
@@ -31,23 +37,24 @@ public class VcsRepositoryFacade {
 
     @Transactional
     public void createVcsRepository(final VcsRepositoryDto vcsRepositoryDto){
-        githubRepositoryApiService.createRepository(vcsRepositoryDto);
-        final VcsRepository vcsRepository = vcsRepositoryMapper.toEntity(vcsRepositoryDto);
-        validateRepository(vcsRepository.getRepository());
+        validateRepository(vcsRepositoryDto.repository());
+        final GithubRepositoryApiResponseDto createdRepository = githubRepositoryApiService.createRepository(vcsRepositoryDto);
+        final VcsRepository vcsRepository = vcsRepositoryMapper.toEntity(createdRepository);
         vcsRepositoryService.saveVcsRepository(vcsRepository);
     }
 
     @Transactional
     public void editVcsRepository(final UUID businessId, final EditVcsRepositoryDto vcsRepositoryDto){
         final VcsRepository vcsRepositoryForUpdate = vcsRepositoryService.getByBusinessId(businessId);
-        githubRepositoryApiService.updateRepository(vcsRepositoryForUpdate.getRepository(), vcsRepositoryDto);
 
         if(ObjectUtils.isNotEmpty(vcsRepositoryDto.repository())
                 && !vcsRepositoryDto.repository().equals(vcsRepositoryForUpdate.getRepository())){
             validateRepository(vcsRepositoryDto.repository());
         }
 
-        final VcsRepository vcsRepository = vcsRepositoryMapper.toEntity(vcsRepositoryDto);
+        final GithubRepositoryApiResponseDto updatedRepository = githubRepositoryApiService.updateRepository(vcsRepositoryForUpdate.getRepository(), vcsRepositoryDto);
+        final VcsRepository vcsRepository = vcsRepositoryMapper.toEntity(updatedRepository);
+
         vcsRepositoryMapper.update(vcsRepositoryForUpdate, vcsRepository);
         vcsRepositoryService.update(vcsRepositoryForUpdate, vcsRepositoryDto.secrets());
     }
@@ -64,13 +71,31 @@ public class VcsRepositoryFacade {
 
     @Transactional
     public void deleteVcsRepository(final UUID businessId) {
-        final VcsRepository repository = vcsRepositoryService.getByBusinessId(businessId);
-        githubRepositoryApiService.deleteRepository(repository.getRepository());
-        vcsRepositoryService.delete(businessId);
+        final ResponseEntity<Void> response = githubRepositoryApiService.deleteRepository(String.valueOf(businessId));
+        if(response.getStatusCode() == HttpStatus.OK){
+            vcsRepositoryService.delete(businessId);
+        } else {
+            throw new GithubException("Failed to delete repository");
+        }
+    }
+
+    public void validateSecret(final VcsRepositoryDto vcsRepositoryDto) {
+        final List<VcsRepositoryView> repositories = listAllVcsRepositories();
+        final boolean isValid = repositories.stream()
+                .filter(repoView -> repoView.repository().equals(vcsRepositoryDto.repository()))
+                .anyMatch(repoView -> repoView.secrets().stream()
+                        .anyMatch(secretView ->
+                                secretView.secret().equals(vcsRepositoryDto.secret().toString())
+                        )
+                );
+
+        if(!isValid){
+            throw new InvalidSecretException();
+        }
     }
 
     private void validateRepository(final String repository){
-        boolean existsByRepository = vcsRepositoryService.doesRepositoryExist(repository);
+        final boolean existsByRepository = vcsRepositoryService.doesRepositoryExist(repository);
         if(existsByRepository){
             throw new DuplicateEntityException();
         }
